+++ app/Http/Controllers/Admin/GroupController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreGroupRequest;
 use App\Http\Requests\UpdateGroupRequest;
 use App\Models\Group;
 use App\Models\Role;
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -15,20 +15,24 @@ class GroupController extends Controller
 {
    
     public function index()
-    {   $groups = Group::search()->latest()->paginate(5);
+    {   
+        $this->authorize('viewAny', Group::class);
+        $groups = Group::search()->latest()->paginate(5);
         return view('admin.groups.index', compact('groups'));
     }
 
   
     public function create()
     {   
+        $this->authorize('create', Group::class);
         $parent_roles = Role::where('group_key', 0)->get();
         return view('admin.groups.create',compact('parent_roles'));
     }
 
   
     public function store(StoreGroupRequest $request)
-    {
+    {   
+        $this->authorize('create', Group::class);
         try {
             $role = Group::create([
                 'name' => $request->name,
@@ -52,7 +56,7 @@ class GroupController extends Controller
     public function edit(string $id)
     {
         try {
-
+            $this->authorize('update', Group::class);
             $group = Group::find($id);
             $roles_checked = $group->roles;
             $parent_roles = Role::where('group_key', 0)->get();
@@ -72,6 +76,7 @@ class GroupController extends Controller
    
     public function update(UpdateGroupRequest $request,$id)
     {
+        $this->authorize('update', Group::class);
         try {
             $group = Group::find($id);
             $group->update([
@@ -87,7 +92,8 @@ class GroupController extends Controller
     }
 
     public function destroy(string $id)
-    {
+    {   
+        $this->authorize('delete', Group::class);
         try {
             $group = Group::find($id);
             DB::table('group_roles')->where('group_id', '=', $group->id)->delete();
